+++ sync/src/download_body.rs
@@ -4,11 +4,9 @@ use actix::prelude::*;
 use anyhow::Result;
 use consensus::Consensus;
 use executor::TransactionExecutor;
-use network::{
-    sync_messages::{DataType, GetDataByHashMsg, ProcessMessage},
-    NetworkAsyncService, RPCRequest, RPCResponse,
-};
+use network::{NetworkAsyncService, RPCRequest, RPCResponse};
 use std::sync::Arc;
+use types::sync_messages::{DataType, GetDataByHashMsg, ProcessMessage};
 use types::{block::BlockHeader, peer_info::PeerInfo};
 
 #[derive(Default, Debug, Message)]
