+++ sync/src/sync.rs
@@ -6,9 +6,9 @@ use bus::{BusActor, Subscription};
 use consensus::Consensus;
 use executor::TransactionExecutor;
 use logger::prelude::*;
-use network::sync_messages::{DownloadMessage, ProcessMessage, SyncMessage};
 use network::PeerEvent;
 use types::peer_info::PeerInfo;
+use types::sync_messages::{DownloadMessage, ProcessMessage, SyncMessage};
 
 pub struct SyncActor<E, C>
 where
