+++ types/Cargo.toml
@@ -37,5 +37,7 @@ libra-types = { git = "https://github.com/starcoinorg/libra", rev="7a35ed6edb792
 libp2p = { git ="https://github.com/starcoinorg/rust-libp2p", branch = "starcoin-merge"}
 logger = {path = "../commons/logger", package="starcoin-logger"}
 move-core-types = {git = "https://github.com/starcoinorg/libra", rev="7a35ed6edb7921ab096028aed91b8e5ff11cc74d" }
+futures = "0.3.4"
+
 
 ethereum-types = "0.2.3"
