+++ types/src/lib.rs
@@ -22,3 +22,5 @@ pub mod transaction;
 pub mod vm_error;
 pub mod write_set;
 pub use ethereum_types::{H256, U256};
+pub mod messages;
+pub mod sync_messages;
